Update Main menu
