+++ client/internal/tui/menu/route.go
@@ -1,0 +1,14 @@
+package menu
+
+type route int
+
+const (
+	routeMenu route = iota
+	routeRegister
+	routeAuth
+	routeSync
+	routeData
+	routeNewItem
+	routeHelp
+	routeQuit
+)
